requirements.txt
pandas==2.0.1
google-cloud-bigquery==3.13.0
bs4
lxml
requests
os